fix(#469): rename the replay mock module and scope its bypass
- rename mocks/node.ts to mocks/server.ts, matching the one-export/file-name
  rule and the client-test-utils precedent
- registerMSWLifecycle takes an optional onUnhandledRequest override; the
  replay preload bypasses only loopback (the ephemeral cross-version provider
  ports) and errors on everything else, instead of a blanket bypass
